child rescue/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-JWT-Extended
python-dotenv
Pillow
face_recognition
numpy
